Get email address from x-amzn-oidc-data header
When using AWS ALB in front of Flagr, we need a way to get the user info
to show in the UI who made changes to a flag. There is no built-in capability
for that so this commit adds it.

If `FLAGR_HEADER_AUTH_USER_FIELD_AWS_ALB` is `true`, we look up the value of
`x-amzn-oidc-data` header which contains 3 base64 encoded strings separated by
dots (`.`). The element at index 1 contains the data we need:
```
{"sub":"00ulryc43ftxDjmqZ0x7","name":"Fardin Khanjani","locale":"DK","email":"fardin.khanjani@tradeshift.com","preferred_username":"fardin.khanjani@tradeshift.com","given_name":"Fardin","family_name":"Khanjani","zoneinfo":"America/Los_Angeles","updated_at":1683740733,"email_verified":true,"exp":1688549942,"iss":"https://tradeshift.okta.com"}
```

we then return the value of `email` attribute.